+++ lib/handler/request_handler.dart
@@ -13,7 +13,8 @@ Future<Client> connect(HttpRequest request) async {
   var token =
       request.headers.value(ConfigService().config.API_ROUTE_AUTH_HEADER);
 
-  var url = ConfigService().config.API_BASE +
+  var url = ConfigService().config.API_SCHEME +
+      ConfigService().config.API_BASE +
       ':' +
       ConfigService().config.API_PORT +
       '/' +